move common logic of highestSupportedVersion to util package
